Bug is fixed.
- Bug is fixed.
- The number of tried schedules is now counted for the brute-force cases.

+++ app/src/main/java/org/cap/model/CoreState.java
@@ -3,7 +3,7 @@ package org.cap.model;
 public class CoreState {
     public Task currentTask;
     public boolean isRunning;
-    public int remainingRuntime;
+    public long remainingRuntime;
     public long minimumVirtualRuntime = 0L;
 
     public CoreState copy() {
